1. buttonIndex类型由int修改为NSInteger  2. 完善注释

+++ LCActionSheetDemo/Classes/LCActionSheet/LCActionSheet.h
@@ -16,7 +16,7 @@
 @optional
 
 /** 点击了buttonIndex处的按钮 */
-- (void)actionSheet:(LCActionSheet *)actionSheet didClickedButtonAtIndex:(int)buttonIndex;
+- (void)actionSheet:(LCActionSheet *)actionSheet didClickedButtonAtIndex:(NSInteger)buttonIndex;
 
 @end
 
@@ -25,7 +25,7 @@
 @property (nonatomic, weak) id<LCActionSheetDelegate> delegate;
 
 /** Tip: 如果没有红色按钮, redButtonIndex给`-1`即可 */
-- (instancetype)initWithTitle:(NSString *)title buttonTitles:(NSArray *)titles redButtonIndex:(int)buttonIndex delegate:(id<LCActionSheetDelegate>)delegate;
+- (instancetype)initWithTitle:(NSString *)title buttonTitles:(NSArray *)titles redButtonIndex:(NSInteger)buttonIndex delegate:(id<LCActionSheetDelegate>)delegate;
 
 - (void)show;
